Extract SpaceService from MessageDB (7 functions)
Extracted 7 space-related functions (~1,140 lines) from MessageDB.tsx into dedicated SpaceService.ts:
- createSpace, updateSpace, deleteSpace
- kickUser, createChannel
- submitUpdateSpace, sendHubMessage

Resolved circular dependencies using forward reference pattern:
- EncryptionService needs updateSpace → SpaceService
- MessageService needs sendHubMessage → SpaceService
- SpaceService needs saveMessage/addMessage → MessageService

Changes:
• Created src/services/SpaceService.ts (1,144 lines)
• Refactored MessageDB.tsx (3,701 → 2,681 lines, -27.5%)
• All 61 automated tests passing
• Updated task tracking documentation
